+++ examples/udp/client.rs
@@ -17,7 +17,7 @@ pub fn run(name: &str) {
     let mut network = NetworkManager::new(move |net_event| network_sender.send(Event::Network(net_event)));
 
     let server_addr = "127.0.0.1:3000";
-    if let Some((server_id, _)) = network.connect_udp(server_addr) {
+    if let Some(server_id) = network.connect_udp(server_addr) {
         println!("Sending to {} by UDP", server_addr);
         event_queue.sender().send(Event::Greet);