+++ src/network.rs
@@ -84,41 +84,53 @@ impl<'a> NetworkManager {
     }
 
     /// Creates a connection to the specific address by TCP.
-    /// The endpoint, an identified of the new connection, will be returned among with the local address of that connection.
+    /// The endpoint, an identified of the new connection, will be returned.
     /// If the connection can not be performed (the address is not reached) a None will be returned.
-    pub fn connect_tcp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<(Endpoint, SocketAddr)> {
-        self.register_connection(Connection::new_tcp_stream(addr.to_socket_addrs().unwrap().next().unwrap()))
+    pub fn connect_tcp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<Endpoint> {
+        let addr = addr.to_socket_addrs().unwrap().next().unwrap();
+        self.register_connection(Connection::new_tcp_stream(addr))
     }
 
     /// Creates a connection to the specific address by UDP.
-    /// The endpoint, an identified of the new connection, will be returned among with the local address of that connection.
+    /// The endpoint, an identified of the new connection, will be returned.
     /// If there is an error during the socket creation, a None will be returned.
-    pub fn connect_udp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<(Endpoint, SocketAddr)> {
-        self.register_connection(Connection::new_udp_socket(addr.to_socket_addrs().unwrap().next().unwrap()))
+    pub fn connect_udp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<Endpoint> {
+        let addr = addr.to_socket_addrs().unwrap().next().unwrap();
+        self.register_connection(Connection::new_udp_socket(addr))
     }
 
     /// Open a port to listen messages from TCP.
     /// If the port can be opened, a endpoint identifying the listener will be returned among with the local address, or a None if not.
     pub fn listen_tcp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<(Endpoint, SocketAddr)> {
-        self.register_connection(Connection::new_tcp_listener(addr.to_socket_addrs().unwrap().next().unwrap()))
+        let addr = addr.to_socket_addrs().unwrap().next().unwrap();
+        self.register_listener(Connection::new_tcp_listener(addr))
     }
 
     /// Open a port to listen messages from UDP.
     /// If the port can be opened, a endpoint identifying the listener will be returned among with the local address, or a None if not.
     pub fn listen_udp<A: ToSocketAddrs>(&mut self, addr: A) -> Option<(Endpoint, SocketAddr)> {
-        self.register_connection(Connection::new_udp_listener(addr.to_socket_addrs().unwrap().next().unwrap()))
+        let addr = addr.to_socket_addrs().unwrap().next().unwrap();
+        self.register_listener(Connection::new_udp_listener(addr))
     }
 
     /// Open a port to listen messages from UDP in multicast.
     /// If the port can be opened, a endpoint identifying the listener will be returned among with the local address, or a None if not.
     pub fn listen_udp_multicast<A: ToSocketAddrs>(&mut self, addr: A) -> Option<(Endpoint, SocketAddr)> {
         match addr.to_socket_addrs().unwrap().next().unwrap() {
-            SocketAddr::V4(addr) => self.register_connection(Connection::new_udp_listener_multicast(addr)),
+            SocketAddr::V4(addr) => self.register_listener(Connection::new_udp_listener_multicast(addr)),
             _ => panic!("listening for udp multicast is only supported for ipv4 addresses"),
         }
     }
 
-    fn register_connection(&mut self, connection: io::Result<Connection>) -> Option<(Endpoint, SocketAddr)> {
+    fn register_connection(&mut self, connection: io::Result<Connection>) -> Option<Endpoint> {
+        connection
+            .ok()
+            .map(|connection| {
+                self.network_controller.add_connection(connection)
+            })
+    }
+
+    fn register_listener(&mut self, connection: io::Result<Connection>) -> Option<(Endpoint, SocketAddr)> {
         connection
             .ok()
             .map(|connection| {